+++ packages/expo-cli/src/commands/push-creds-web.ts
@@ -17,8 +17,9 @@ const vapidSubjectDescription =
 
 export default function (program: Command) {
   program
-    .command('push:web:upload [project-dir]')
-    .description('Uploads VAPID key pair and VAPID subject for web push notifications.')
+    .command('push:web:upload [path]')
+    .description('Upload VAPID key pair and VAPID subject for web push notifications.')
+    .helpGroup('notifications')
     .option('--vapid-pubkey [vapid-public-key]', 'URL-safe base64-encoded VAPID public key.')
     .option('--vapid-pvtkey [vapid-private-key]', 'URL-safe base64-encoded VAPID private key.')
     .option('--vapid-subject [vapid-subject]', vapidSubjectDescription)
@@ -33,8 +34,9 @@ export default function (program: Command) {
     });
 
   program
-    .command('push:web:generate [project-dir]')
-    .description('Generates VAPID key pair for web push notifications.')
+    .command('push:web:generate [path]')
+    .description('Generate VAPID key pair for web push notifications.')
+    .helpGroup('notifications')
     .option('--vapid-subject [vapid-subject]', vapidSubjectDescription)
     .asyncActionProjectDir(async (projectDir: string, options: VapidData) => {
       if (!options.vapidSubject) {
@@ -47,10 +49,11 @@ export default function (program: Command) {
     });
 
   program
-    .command('push:web:show [project-dir]')
+    .command('push:web:show [path]')
     .description(
-      'Prints the VAPID public key, the VAPID private key, and the VAPID subject currently in use for web notifications for this project.'
+      'Log the VAPID public key, the VAPID private key, and the VAPID subject currently in use for web notifications for this project.'
     )
+    .helpGroup('notifications')
     .asyncActionProjectDir(async (projectDir: string) => {
       const {
         args: { remotePackageName },
@@ -75,10 +78,11 @@ export default function (program: Command) {
     });
 
   program
-    .command('push:web:clear [project-dir]')
+    .command('push:web:clear [path]')
     .description(
-      'Deletes previously uploaded VAPID public key, VAPID private key, and VAPID subject.'
+      'Delete previously uploaded VAPID public key, VAPID private key, and VAPID subject.'
     )
+    .helpGroup('notifications')
     .asyncActionProjectDir(async (projectDir: string) => {
       log('Reading project configuration...');
       const {