+++ packages/expo-cli/src/commands/register.ts
@@ -5,6 +5,7 @@ import { register } from '../accounts';
 export default function (program: Command) {
   program
     .command('register')
+    .helpGroup('auth')
     .description('Sign up for a new Expo account')
     .asyncAction(() => register());
 }